% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpois_custom.R
\name{qpois_custom}
\alias{qpois_custom}
\title{Poisson Distribution: Quantile}
\usage{
qpois_custom(p, lambda, lower.tail = TRUE)
}
\arguments{
\item{p}{Numeric, target probability (must be between 0 and 1).}

\item{lambda}{Numeric, parameter of the Poisson distribution (must be > 0).}

\item{lower.tail}{Logical; if TRUE, returns the quantile for P(X ≤ x) ≥ p, otherwise for P(X > x) ≤ p.}
}
\value{
The quantile as an integer.
}
\description{
Calculates the quantile for a given probability according to a Poisson distribution.
}
\examples{
# Quantile for a cumulative probability of 0.8 with lambda = 3
qpois_custom(0.8, 3)
# Returns: 4

# Quantile for a complementary probability of 0.8 with lambda = 3
qpois_custom(0.8, 3, lower.tail = FALSE)
# Returns: 1
}
