% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpois_custom.R
\name{rpois_custom}
\alias{rpois_custom}
\title{Poisson Distribution: Random Generation}
\usage{
rpois_custom(n, lambda)
}
\arguments{
\item{n}{Integer, number of random values to generate (must be > 0).}

\item{lambda}{Numeric, parameter of the Poisson distribution (must be > 0).}
}
\value{
A vector of n integer values.
}
\description{
Generates random values following a Poisson distribution for a given lambda parameter.
}
\examples{
# Generates 10 values following a Poisson distribution with lambda = 3
rpois_custom(10, 3)
# Example output: 3, 2, 1, 5, 4, 2, 3, 3, 4, 1
}
