% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_poisson_fit.R
\name{test_poisson_fit}
\alias{test_poisson_fit}
\title{Goodness-of-Fit Test for the Poisson Distribution}
\usage{
test_poisson_fit(data, lambda)
}
\arguments{
\item{data}{Vector of observed data, representing the number of occurrences of an event in given time intervals.}

\item{lambda}{Hypothetical parameter of the Poisson distribution (> 0), representing the expected average number of events.}
}
\value{
An object of type `htest` with the results of the chi-squared test, including the test statistic and the p-value.
}
\description{
This function performs a chi-squared test to check if a series of data follows a Poisson distribution with a given lambda parameter.
}
\examples{
data <- rpois(50, lambda = 3)
test_poisson_fit(data, lambda = 3)
}
